+++ _output/extension_hint.php
@@ -39,6 +39,7 @@ namespace Jrahmy\Moderation\XF\Repository
 namespace Jrahmy\Moderation\XF\Service\Report
 {
 	class XFCP_Commenter extends \XF\Service\Report\Commenter {}
+	class XFCP_CommentPreparer extends \XF\Service\Report\CommentPreparer {}
 	class XFCP_Creator extends \XF\Service\Report\Creator {}
 	class XFCP_Notifier extends \XF\Service\Report\Notifier {}
 }